Refacto unit tests

See merge request tkint/MiniDAO!12

+++ src/test/java/com/thomaskint/minidao/querybuilder/MDInsertBuilderTest.java
@@ -43,8 +43,8 @@ public class MDInsertBuilderTest {
 	public void should_construct_query() {
 		// GIVEN
 		String expectedQuery = "INSERT INTO message (id_author, content) VALUES (1, 'test-content')";
-		UserTest author = new UserTest(BigDecimal.valueOf(1), "pseudo", "login", "password");
-		MessageTest message = new MessageTest(BigDecimal.valueOf(1), BigDecimal.valueOf(1), author, "test-content");
+		UserTest author = new UserTest(1L, "pseudo", "login", "password");
+		MessageTest message = new MessageTest(1L, 1L, author, "test-content");
 		MDInsertBuilder insertBuilder = new MDInsertBuilder(MiniDAO.getDefaultDateFormat());
 		// WHEN
 		String returnedQuery = null;
